& + double

+++ GeoLibRefactor/include/Triangle.h
@@ -89,9 +89,9 @@ namespace GeoLib {
          */
         vector3 computeVoronoi(vector3 a, vector3 b, vector3 c);
 
-        float computeAspectRation(vector3 a, vector3 b, vector3 c);
+        double computeAspectRation(vector3 a, vector3 b, vector3 c);
 
-        float minAngle(vector3 a, vector3 b, vector3 c);
+        double minAngle(vector3 a, vector3 b, vector3 c);
 
     private:
         bool updateVoronoi = true;
